Implement interactive path updates
Add first draft of the animation of the paths of the citizens along their way.

The scale of second used changed from 1000:1 to 100:1 as any time value below 10 ms will be wrapped to 10 ms causing unrealisitic behavior of the animation. By increasing this value to 100:1 time values below 10 ms are avoided.