Implement feature that selects most fastest route
As before, it was possible to get a ride that took more time because of the travel time to get to the vehicle. Now it picks the route with the vehicle in case it is indeed faster.

Change the jQuery ajax for superagent for Promises support. Add Bluebird's Promises for the Promise.map function.